Refactor UI specs to use installSpotifyRoutes and restore requested tests
